sepai-backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
httpx      # optional, used for some client calls
python-jose[cryptography]  # optional, for JWT verification if needed
